+++ src/main/java/web/service/UserService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface UserService {
     void saveUser(User user);
 
-    void updateUser(int id,User user);
+    void updateUser(long id,User user);
 
-    void deleteUser(int id);
+    void deleteUser(long id);
 
     List<User> getAllUsers();
 
-    User showUser(int id);
+    User showUser(long id);
 }
